+++ src/include/ci/efch/mmap.h
@@ -28,7 +28,7 @@ efab_iobufset_resource_mmap(struct iobufset_resource *iobrs,
                             unsigned long* bytes, void* opaque, int* map_num,
                             unsigned long* offset, int index);
 
-extern unsigned
+extern unsigned long
 efab_iobufset_resource_nopage(struct iobufset_resource* iobrs, void* opaque, 
                               unsigned long offset, unsigned long map_size);
 
@@ -40,7 +40,7 @@ extern int
 efab_vi_resource_mmap_bytes(struct efrm_vi* virs, int map_type);
 
 #ifdef CI_HAVE_OS_NOPAGE
-extern unsigned
+extern unsigned long
 efab_vi_resource_nopage(struct efrm_vi *virs, void *opaque,
                         unsigned long offset, unsigned long map_size);
 #endif
